+++ tests/it/mixed.rs
@@ -1,27 +1,18 @@
 use unimock::*;
 
+mod clone {
+
+    #[derive(Eq, PartialEq, Debug)]
+    pub struct Nope;
+
+    #[derive(Clone, Eq, PartialEq, Debug)]
+    pub struct Sure;
+}
+
 #[unimock(api=InOptionMock)]
 trait InOption {
     fn str(&self, a: &str) -> Option<&str>;
-}
-
-#[derive(Eq, PartialEq, Debug)]
-pub struct Error;
-
-#[derive(Clone, Eq, PartialEq, Debug)]
-pub struct ErrorClone;
-
-#[unimock(api=InResultMock)]
-trait InResult {
-    fn bytes(&self) -> Result<&[u8], Error>;
-    fn u32_clone(&self) -> Result<&u32, ErrorClone>;
-}
-
-// Note: This should only be mockable with static lifetimes
-// It should not pick a Mixed output mediator.
-#[unimock]
-trait InResultWithComplexLifetimes {
-    fn foo<'s, 'i>(&'s self, a: &'i str) -> Result<&'s str, &'i str>;
+    fn not_clone(&self) -> Option<&clone::Nope>;
 }
 
 #[test]
@@ -36,6 +27,35 @@ fn in_option() {
     assert_eq!(Some("2"), <Unimock as InOption>::str(&u, ""));
 }
 
+// This test demonstrates that a `T: Clone` bound is not necessary
+// on each_call for `Option<&T>`, since the outer option
+// can be generically reconstructed from the inner borrowed one
+#[test]
+fn in_option_not_clone_can_clone_anyway() {
+    let u = Unimock::new(
+        InOptionMock::not_clone
+            .each_call(matching!())
+            .returns(Some(clone::Nope)),
+    );
+    assert_eq!(Some(&clone::Nope), <Unimock as InOption>::not_clone(&u));
+    assert_eq!(Some(&clone::Nope), <Unimock as InOption>::not_clone(&u));
+}
+
+#[unimock(api=InResultMock)]
+trait InResult {
+    fn bytes(&self) -> Result<&[u8], clone::Nope>;
+    fn u32_clone(&self) -> Result<&u32, clone::Sure>;
+    fn ok_no_clone(&self, ok: bool) -> Result<&clone::Nope, clone::Sure>;
+    fn err_no_clone(&self) -> Result<&clone::Nope, clone::Nope>;
+}
+
+// Note: This should only be mockable with static lifetimes
+// It should not pick a Mixed output mediator.
+#[unimock]
+trait InResultWithComplexLifetimes {
+    fn foo<'s, 'i>(&'s self, a: &'i str) -> Result<&'s str, &'i str>;
+}
+
 #[test]
 fn in_result() {
     let u = Unimock::new((
@@ -44,7 +64,7 @@ fn in_result() {
             .returns(Ok(vec![42])),
         InResultMock::bytes
             .next_call(matching!())
-            .returns(Result::<&[u8], _>::Err(Error)),
+            .returns(Err::<&[u8], _>(clone::Nope)),
         InResultMock::u32_clone
             .each_call(matching!())
             .returns(Ok(42))
@@ -52,7 +72,44 @@ fn in_result() {
     ));
 
     assert_eq!(Ok(vec![42].as_slice()), <Unimock as InResult>::bytes(&u));
-    assert_eq!(Err(Error), <Unimock as InResult>::bytes(&u));
+    assert_eq!(Err(clone::Nope), <Unimock as InResult>::bytes(&u));
     assert_eq!(Ok(&42), <Unimock as InResult>::u32_clone(&u));
     assert_eq!(Ok(&42), <Unimock as InResult>::u32_clone(&u));
+}
+
+#[test]
+fn in_result_clone_acrobatics() {
+    let u = Unimock::new((
+        InResultMock::ok_no_clone
+            .each_call(matching!(true))
+            .returns(Ok(clone::Nope)),
+        InResultMock::ok_no_clone
+            .each_call(matching!(false))
+            .returns(Err::<&clone::Nope, _>(clone::Sure)),
+        InResultMock::err_no_clone
+            .some_call(matching!()) // note: .each_call is impossible
+            .returns(Err::<&clone::Nope, _>(clone::Nope)),
+    ));
+
+    for _ in 0..3 {
+        assert_eq!(Ok(&clone::Nope), u.ok_no_clone(true));
+        assert_eq!(Err(clone::Sure), u.ok_no_clone(false));
+    }
+
+    assert_eq!(Err(clone::Nope), u.err_no_clone());
+}
+
+#[test]
+#[should_panic(
+    expected = "InResult::ok_no_clone: Expected InResult::ok_no_clone(_) at tests/it/mixed.rs:109 to match exactly 1 call, but it actually matched 2 calls."
+)]
+fn in_result_may_multi_respond_on_ok_no_clone() {
+    let u = Unimock::new(
+        InResultMock::ok_no_clone
+            .some_call(matching!(_))
+            .returns(Ok(clone::Nope)),
+    );
+
+    assert_eq!(Ok(&clone::Nope), u.ok_no_clone(true));
+    assert_eq!(Ok(&clone::Nope), u.ok_no_clone(true));
 }